refactor: Order Domain Context > Order, OrderLineItem, OrderStatus - 패키지 생성 및 기존 클래스 이동
