+++ src/main/java/kitchenpos/dao/OrderLineItemDao.java
@@ -2,7 +2,7 @@ package kitchenpos.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public interface OrderLineItemDao {
     OrderLineItem save(OrderLineItem entity);
