+++ src/test/java/kitchenpos/fixture/OrderFixtureFactory.java
@@ -1,9 +1,9 @@
 package kitchenpos.fixture;
 
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 
 public class OrderFixtureFactory {
     private OrderFixtureFactory() {
