+++ src/test/java/kitchenpos/fixture/OrderLineItemFixtureFactory.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemFixtureFactory {
     private OrderLineItemFixtureFactory() {
