+++ src/test/java/kitchenpos/utils/ServiceTestHelper.java
@@ -5,12 +5,12 @@ import static java.util.stream.Collectors.toList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
-import kitchenpos.application.OrderService;
+import kitchenpos.order.application.OrderService;
 import kitchenpos.table.application.TableGroupService;
 import kitchenpos.table.application.TableService;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.fixture.MenuFixtureFactory;
 import kitchenpos.fixture.MenuGroupFixtureFactory;
 import kitchenpos.fixture.OrderFixtureFactory;
